# apps/api/src/public/swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getData","parameters":[],"responses":{"200":{"description":""}}}},"/api/gateways":{"post":{"operationId":"GatewayController_createGateway","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGatewayDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"code":{"type":"number"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/Gateway"}}}]}}}}}},"get":{"operationId":"GatewayController_getAllGateways","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":15,"type":"number"}},{"name":"searchKey","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"code":{"type":"number"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Gateway"}},"pageInfo":{"$ref":"#/components/schemas/PaginationInfo"}}}}}]}}}}}}},"/api/gateways/{name}":{"get":{"operationId":"GatewayController_getGateway","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"code":{"type":"number"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/Gateway"}}}]}}}}}}},"/api/gateways/{id}/devices":{"post":{"operationId":"GatewayController_addDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePeripheralDeviceDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"code":{"type":"number"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PeripheralDevice"}}}]}}}}}}},"/api/gateways/{id}/devices/{deviceUId}":{"delete":{"operationId":"GatewayController_removeDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"deviceUId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"code":{"type":"number"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/Boolean"}}}]}}}}}}}},"info":{"title":"Swagger Api","description":"Swagger Api Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ApiResponse":{"type":"object","properties":{"code":{"type":"number"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object"}},"required":["code","success","message","data"]},"PaginationInfo":{"type":"object","properties":{"page":{"type":"number"},"limit":{"type":"number"},"totalCount":{"type":"number"},"hasNext":{"type":"boolean"},"hasBefore":{"type":"boolean"}},"required":["hasNext","hasBefore"]},"PaginatedResponse":{"type":"object","properties":{"items":{"type":"array","items":{"type":"array"}},"pageInfo":{"$ref":"#/components/schemas/PaginationInfo"}},"required":["items","pageInfo"]},"PeripheralDevice":{"type":"object","properties":{"uid":{"type":"number"},"vendor":{"type":"string"},"status":{"type":"string","enum":["online","offline"]},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["uid","vendor","status","createdAt","updatedAt"]},"Gateway":{"type":"object","properties":{"serialNumber":{"type":"string"},"name":{"type":"string"},"ipv4":{"type":"string"},"devices":{"type":"array","items":{"$ref":"#/components/schemas/PeripheralDevice"}}},"required":["serialNumber","name","ipv4","devices"]},"CreatePeripheralDeviceDto":{"type":"object","properties":{"vendor":{"type":"string","description":"Vendor of the peripheral device"},"status":{"type":"string","description":"Status of the peripheral device","enum":["online","offline"]}},"required":["vendor","status"]},"CreateGatewayDto":{"type":"object","properties":{"serialNumber":{"type":"string","description":"Unique serial number of the gateway"},"name":{"type":"string","description":"Human-readable name of the gateway"},"ipv4":{"type":"string","description":"IPv4 address of the gateway","example":"192.168.1.1"},"devices":{"description":"List of peripheral devices associated with the gateway","type":"array","items":{"$ref":"#/components/schemas/CreatePeripheralDeviceDto"}}},"required":["serialNumber","name","ipv4","devices"]},"Boolean":{"type":"object","properties":{}}}}}